ec/curve25519.c: fe51_mul IFMA proof of concept.
General idea is to calculate h[0:3] to %ymm registers using vpmadd52
instructions, and h[4] - in scalar. As name suggests, vpmadd52[lh]uq
use 52 bits of their inputs, which is why you'll spot 'and' MASK52.
While they might appear redundant, they serve as kind of assertions
that no bits run unaccounted for. Rationale is that if some significant
bits are zeroed, then tests are likely to fail.